docs(auto-chart): basic demo and zh api doc
