+++ sites/ava-site/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  rules: {
+    // to handle gatsby browser defined local packages not error
+    'import/no-unresolved': 0,
+  }
+};
