+++ sites/ava-site/examples/auto-chart/API.en.md
@@ -1,0 +1,1 @@
+`markdown:docs/api/autoChart.en.md`
