+++ sites/ava-site/examples/auto-chart/API.zh.md
@@ -1,0 +1,1 @@
+`markdown:docs/api/autoChart.zh.md`
