+++ sites/ava-site/examples/auto-chart/demo/meta.json
@@ -1,0 +1,24 @@
+{
+  "title": {
+    "zh": "中文分类",
+    "en": "Category"
+  },
+  "demos": [
+    {
+      "filename": "basic.jsx",
+      "title": {
+        "en": "Basic autoChart usage",
+        "zh": "基础 AutoChart 用法"
+      },
+      "screenshot": "https://gw.alipayobjects.com/mdn/rms_fabca5/afts/img/A*hbeeS6gtKVQAAAAAAAAAAAAAARQnAQ"
+    },
+    {
+      "filename": "mock.jsx",
+      "title": {
+        "en": "AutoChart mock data",
+        "zh": "AutoChart 无数据状态"
+      },
+      "screenshot": "https://gw.alipayobjects.com/mdn/rms_fabca5/afts/img/A*hbeeS6gtKVQAAAAAAAAAAAAAARQnAQ"
+    }
+  ]
+}
