+++ sites/ava-site/examples/auto-chart/demo/mock.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { AutoChart } from '@antv/auto-chart';
+
+ReactDOM.render(<AutoChart />, document.getElementById('container'));
