+++ sites/ava-site/examples/auto-chart/index.en.md
@@ -1,5 +1,5 @@
 ---
-title: autoChart
+title: AutoChart
 order: 0
 ---
