+++ sites/ava-site/examples/auto-chart/index.zh.md
@@ -1,0 +1,4 @@
+---
+title: AutoChart 示例
+order: 0
+---
