--- sites/ava-site/examples/chart-advisor/auto-chart/API.en.md
@@ -1,3 +1,0 @@
-## autoChart
-
-`markdown:docs/api/chart-advisor/autoChart.en.md`
