--- sites/ava-site/examples/chart-advisor/auto-chart/demo/meta.json
@@ -1,32 +1,0 @@
-{
-  "title": {
-    "zh": "中文分类",
-    "en": "Category"
-  },
-  "demos": [
-    {
-      "filename": "basic.ts",
-      "title": {
-        "en": "Basic autoChart usage",
-        "zh": "基础 autoChart 用法"
-      },
-      "screenshot": "https://gw.alipayobjects.com/mdn/rms_fabca5/afts/img/A*hbeeS6gtKVQAAAAAAAAAAAAAARQnAQ"
-    },
-    {
-      "filename": "refine.ts",
-      "title": {
-        "en": "Refine Chart",
-        "zh": "开启图表细节优化"
-      },
-      "screenshot": "https://gw.alipayobjects.com/mdn/rms_fabca5/afts/img/A*ybCXT4O_DXwAAAAAAAAAAAAAARQnAQ"
-    },
-    {
-      "filename": "custom-rules.jsx",
-      "title": {
-        "en": "Customized Rules",
-        "zh": "自定义推荐规则"
-      },
-      "screenshot": "https://gw.alipayobjects.com/zos/antfincdn/QNOIESEqnK/customrules.png"
-    }
-  ]
-}
