--- sites/ava-site/examples/chart-advisor/auto-chart/index.zh.md
@@ -1,4 +1,0 @@
----
-title: autoChart 示例
-order: 0
----
